[#6] Separate out Familia::VERSION
